requirements.txt
opencv-python
EasyOCR
torch
colorama
Pillow
pymongo
python-dotenv
